backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
scipy
python-multipart
google-generativeai
